+++ library/core/src/main/java/com/google/android/exoplayer2/source/ClippingMediaPeriod.java
@@ -211,6 +211,11 @@ public final class ClippingMediaPeriod implements MediaPeriod, MediaPeriod.Callb
     return mediaPeriod.continueLoading(positionUs);
   }
 
+  @Override
+  public boolean isLoading() {
+    return mediaPeriod.isLoading();
+  }
+
   // MediaPeriod.Callback implementation.
 
   @Override