+++ library/hls/src/main/java/com/google/android/exoplayer2/source/hls/HlsSampleStreamWrapper.java
@@ -664,6 +664,11 @@ import java.util.Set;
   }
 
   @Override
+  public boolean isLoading() {
+    return loader.isLoading();
+  }
+
+  @Override
   public void reevaluateBuffer(long positionUs) {
     // Do nothing.
   }